pyproject.toml
[project]
name = "transcript-gen"
version = "0.1.0"
description = "Automated transcript generation for monitor evasion research"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.6",
    "anthropic>=0.42",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.scripts]
transcript-gen = "transcript_gen.optimize:main"
transcript-eval = "transcript_gen.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcript_gen"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
